acoplate hashManager
