--- semana20/src/service/hashManager.test.ts
@@ -1,17 +1,0 @@
-import { compare, hash } from "./hashManager"
-
-describe('hashManager', () => {
-    const plaintext = 'secret'
-    
-    test('hash exist', async () => {
-        const cypherText = await hash(plaintext)
-        expect(typeof cypherText).toBe('string')
-    })
-
-    test('decrypt', async () => {
-        const cypherText = await hash(plaintext)
-        expect( await compare(plaintext, cypherText) )
-        .toBe(true)
-    })
-
-})
